+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - v? (?):
+    + Add JSON output to 'list' command.
     + ...
 
 - v0.3.8 (2019-12-26):